Moved actions.js to own folder in preparation of splitting it up.
